+++ NEWS.rst
@@ -2,6 +2,14 @@
 Buildthings Releases
 ====================
 
-Buildthings 1.0 (27-March-2026):
+Buildthings 1.0.1 (TBD)
+=======================
+
+* Existing symlinks in ``.npm-workspaces`` are now restored after the build
+  completes or fails.
+
+
+Buildthings 1.0 (27-March-2026)
+===============================
 
 * Initial public release.